+++ usian-leadnews-apis/src/main/java/com/usian/api/wemedia/WmUserControllerApi.java
@@ -26,6 +26,6 @@ public interface WmUserControllerApi {
      * @param id
      * @return
      */
-    WmUser findWmUserById(Long id);
+    WmUser findWmUserById(Integer id);
 
 }
