+++ usian-leadnews-wemedia/src/main/java/com/usian/wemedia/controller/v1/WmUserController.java
@@ -37,7 +37,7 @@ public class WmUserController implements WmUserControllerApi {
 
     @GetMapping("/findOne/{id}")
     @Override
-    public WmUser findWmUserById(@PathVariable("id") Long id) {
+    public WmUser findWmUserById(@PathVariable("id") Integer id) {
         return userService.getById(id);
     }
 }
